tests VAE
