+++ code/similarity_learning/models/vae/unitTest/VAETest.py
@@ -335,6 +335,7 @@ class TestVAEFunctions(unittest.TestCase):
         epoch_nb = 11
         datasetName = 'dummyDataset98.npz'
         datasetDir = './dummyDataset/'
+        saveDir = './dummySaveTest/'
         testDataset = NPZ_Dataset(datasetName,
                                   datasetDir, 'Spectrums')
         train_loader = torch.utils.data.DataLoader(
@@ -353,14 +354,14 @@ class TestVAEFunctions(unittest.TestCase):
         vae.trainVAE(train_loader, epoch_nb)
         # save it
         if vae.trained:
-            vae.save(datasetName, datasetDir)
+            vae.save(datasetName, saveDir)
 
         self.assertTrue(vae.created and
                         vae.trained and vae.saved)
 
     #---------------------------------------
 
-    def test_gaussianVAE_trainsaveload(self):
+    def test_gaussianVAE_trainsave(self):
         mb_size = 49  # because dummyDataset98.npz is a 98 data size
         epoch_nb = 5
         # if exists remove 'saveloadTest' folder
@@ -388,17 +389,16 @@ class TestVAEFunctions(unittest.TestCase):
         # save it
         savefile = vae.save(datasetName, saveDir)
         # reload the savefile of VAE
-        vae = loadVAE(savefile, saveDir)
+        # vae = loadVAE(savefile, saveDir)
 
         # continue training
         # vae.trainVAE(train_loader, 10)
-        # vae.save(datasetName, datasetDir)
-        self.assertTrue(vae.created and vae.loaded and vae.saved)
+        # vae.save(datasetName, saveDir)
+        self.assertTrue(vae.created and vae.saved) #and vae.loaded
 
     def test_VAE_load(self):
-        # try to load a vae
-        vae = loadVAE('dummyDataset98_NPZ_E<1024-relu6-600-muSig-10>\
-                    _D<10-relu6-600-muSig-1024>_beta1_mb49_lr0dot001_ep5',
+        # try to load a vae 
+        vae = loadVAE('dummyDataset98_NPZ_E<1024-relu6-600-muSig-10>_D<10-relu6-600-muSig-1024>_beta1_mb49_lr0dot001_ep5',
                       './dummySaveTest/')
         self.assertTrue(vae.created and vae.loaded)
 
@@ -426,8 +426,7 @@ class TestVAEVisualize(unittest.TestCase):
 
     def test_VAE_PCA(self):
         # try to load a vae
-        vaeLoaded = loadVAE('dummyDataset98_NPZ_E<1024-relu6-600-muSig-10>_D<10-relu6-600-muSig-1024>_beta1_mb49_lr0dot001_ep5',
-                            './dummySaveTest/')
+        vaeLoaded = loadVAE('dummyDataset98_NPZ_E<1024-relu6-401-muSig-6>_D<6-relu6-399-sigmoid-1024>_beta1_mb49_lr0dot001_ep11','./dummySaveTest/')
         test = numpy.load("./dummyDataset/dummyDataset98.npz")
         # vaeLoaded = loadVAE('toy-spectral-richness-v2-lin_NPZ_E<1024-relu6-600-muSig-10>_D<10-relu6-600-muSig-1024>_beta3_mb100_lr0dot001_ep400',
         #                     "../data/savedVAE/Alexis-nonNorm/beta3/WU100/")
@@ -438,7 +437,7 @@ class TestVAEVisualize(unittest.TestCase):
         lbltest = []
         lbltest = lbl[:, 0:98:]
         spec = spec[:, 0:98]
-        print(len(lbltest[0, :]))
+        print("~~~ Nb samples :", len(lbltest[0, :]))
 
         spec = torch.from_numpy(spec)
         spec = spec.float()
